core/linux-nyan Fix remaining GCC6 issues
GCC 6 is way stricter than previous versions and will warn about too many
things. But all warnings are treated as errors when building this package,
so even stylistic things such as identation of the code result in the compiler
bailing out with an error. This commit instructs the compiler to not treat
unused-const-variables, misleading-indentations, and possible array-bounds
violations as errors.